Continued work on adding support for more floating point types

+++ source/tools/finite/opt_meta.h
@@ -26,9 +26,9 @@ namespace tools::finite::opt {
         size_t              bondexp_maxiter   = 1ul;
         size_t              bondexp_blocksize = 1ul;
         size_t              bondexp_nkrylov   = 10ul;
-        double              bondexp_factor    = 1.0;
-        double              bondexp_minalpha  = 1e-15;
-        double              bondexp_maxalpha  = 1e-3;
+        float               bondexp_factor    = 1.0f;
+        float               bondexp_minalpha  = 1e-15f;
+        float               bondexp_maxalpha  = 1e-3f;
         size_t              max_sites         = 2ul;
         size_t              min_sites         = 1ul;
         long                max_problem_size  = 0;
